+++ src/main/java/zone/moddev/mc/skysgrassslabs/SkysGrassSlabs.java
@@ -18,7 +18,7 @@ import zone.moddev.mc.skysgrassslabs.world.GrassSlabSmoothingFeature;
 public final class SkysGrassSlabs {
     public static final String MOD_ID = "skysgrassslabs";
     public static final String NAME = "Sky's Grass Slabs";
-    public static final String VERSION = "1.0.1.114041";
+    public static final String VERSION = "1.0.1.115021";
     public static final Logger LOGGER = LogManager.getLogger();
 
     public SkysGrassSlabs() {