+++ src/main/java/zone/moddev/mc/skysgrassslabs/block/TurfBlock.java
@@ -6,9 +6,7 @@ import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
-import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.Direction;
-import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
@@ -16,6 +14,7 @@ import net.minecraft.world.IBlockReader;
 import net.minecraft.world.IWorld;
 import net.minecraft.world.IWorldReader;
 import net.minecraft.world.World;
+import net.minecraft.world.server.ServerWorld;
 
 public final class TurfBlock extends Block {
     public static final VoxelShape TURF_SHAPE =
@@ -56,10 +55,7 @@ public final class TurfBlock extends Block {
     }
 
     @Override
-    public void tick(BlockState state, World world, BlockPos pos, Random random) {
-        if (world.isRemote) {
-            return;
-        }
+    public void tick(BlockState state, ServerWorld world, BlockPos pos, Random random) {
         BlockPos support = pos.down();
         if (world.getBlockState(support).getBlock() != Blocks.DIRT) {
             world.destroyBlock(pos, true);
@@ -78,11 +74,6 @@ public final class TurfBlock extends Block {
     public int getFireSpreadSpeed(BlockState state, IBlockReader world, BlockPos pos,
             Direction face) {
         return 30;
-    }
-
-    @Override
-    public BlockRenderLayer getRenderLayer() {
-        return BlockRenderLayer.CUTOUT_MIPPED;
     }
 
     private static boolean hasFullSupport(IWorldReader world, BlockPos pos) {